refactoring subscription form
